Rename to pac-boussole
Big rename all over to call it pac boussole
