+++ .github/workflows/container.yaml
@@ -1,6 +1,6 @@
 name: Build Multi-Arch Image
 env:
-  IMAGE_NAME: pipelines-as-code-prow
+  IMAGE_NAME: pac-boussole
   IMAGE_TAG: nightly
   IMAGE_REGISTRY: ghcr.io/${{ github.repository_owner }}
   REGISTRY_USER: ${{ github.actor }}
