+++ .tekton/pipelines/in-repo.yaml
@@ -1,5 +1,3 @@
-# Pipeline using the pac-pro code from the revision where it started
----
 apiVersion: tekton.dev/v1
 kind: Pipeline
 metadata:
@@ -46,7 +44,7 @@ spec:
               - name: revision
                 value: "$(params.revision)"
           - name: manage-pr
-            image: ghcr.io/openshift-pipelines/pipelines-as-code-prow:nightly
+            image: ghcr.io/openshift-pipelines/pac-boussole:nightly
             workingDir: $(workspaces.source.path)
             env:
               - name: GITHUB_TOKEN
@@ -76,6 +74,6 @@ spec:
               - name: PAC_LGTM_REVIEW_EVENT
                 value: $(params.lgtm_review_event)
             script: |
-              exec ./pac-prow
+              exec ./pac-boussole
       workspaces:
         - name: source
