+++ pyproject.toml
@@ -1,16 +1,10 @@
 [project]
-name = "pipelines-as-code-prow"
+name = "pac-boussole"
 version = "0.1.0"
-description = "Pipelines as Code prow let you interact with tekton pipelines via Gitops command a la prow."
+description = "A GitOps command tool that manages GitHub Pull Requests with Kubernetes Prow like commands"
 readme = "README.md"
 requires-python = ">=3.12"
-dependencies = [
-    "requests>=2.32.3",
-]
+dependencies = ["requests>=2.32.3"]
 
 [dependency-groups]
-dev = [
-    "pre-commit>=4.1.0",
-    "pylint>=3.3.4",
-    "pytest",
-]
+dev = ["pre-commit>=4.1.0", "pylint>=3.3.4", "pytest"]
